requirements.txt
fastapi
uvicorn[standard]
pypdf2
pymupdf
easyocr
pillow
jinja2
python-multipart
